navigation sliding issues
