+++ screens/ChatListScreen/index.js
@@ -22,36 +22,7 @@ import BBBIcon from '../../components/BBBIcon';
 import styles from './styles';
 import { Layout, Colors, Images } from '../../constants/';
 
-//apollo client
-import { Query } from "react-apollo";
-import gql from "graphql-tag";
-import { ApolloClient } from "apollo-client";
-import { HttpLink } from "apollo-link-http";
-import { InMemoryCache } from "apollo-cache-inmemory";
-import { ApolloProvider, graphql,Mutation } from "react-apollo";
-import { withClientState } from "apollo-link-state";
 
-// Get login status
-var log_status = '';
-const GET_LOGIN_STATUS = gql`
-     query log @client{
-           logged_in
-        }`;
-
-const App = () => (
-<Query query={GET_LOGIN_STATUS}>
-  {({ loading, error, data }) => {
-     if (loading) return <Text>{`Loading...`}</Text>;
-     if (error) return <Text>{`Error: ${error}`}</Text>;
-      console.log('get data');
-      console.log('profile_query '+data.logged_in);
-      log_status = data.logged_in;
-    return (
-      <View/>
-    )
-  }}
-</Query>
-)
 
 export default class ChatListScreen extends React.Component {
 
@@ -61,18 +32,6 @@ export default class ChatListScreen extends React.Component {
 			data: [],
 		}
 	}
-
-
-// Check weather user is login or not
-componentDidMount = async () => {
-
-				console.log("Log Status: " + log_status);
-
-				if(log_status == false)
-				{
-					this.props.navigation.navigate('loginScreen');
-				}
-}
 
 
 	_renderItem = ({ item }) => (
@@ -182,7 +141,7 @@ componentDidMount = async () => {
 			<Container style={styles.container}>
 				<BBBHeader title="Chats" leftComponent={leftComponent} />
 				<Content>
-					<App/>
+				
 					<FlatList
 						data={listItemData}
 						keyExtractor={listItemData => listItemData.id}
